refactor:#1 스토리북 preview 파일 수정
