+++ .storybook/preview.tsx
@@ -1,0 +1,26 @@
+import React from 'react';
+
+import type { Preview } from '@storybook/react';
+import { ThemeProvider } from 'styled-components';
+import theme from 'src/styles/theme';
+
+const preview: Preview = {
+  parameters: {
+    controls: {
+      matchers: {
+        color: /(background|color)$/i,
+        date: /Date$/i,
+      },
+    },
+  },
+};
+
+export const decorators = [
+  (Story) => (
+    <ThemeProvider theme={theme}>
+      <Story />
+    </ThemeProvider>
+  ),
+];
+
+export default preview;
